feature: resolve merge issues
