+++ internal/admin-auth/handler/tests/handler_oauth_test.go
@@ -2,9 +2,7 @@ package tests
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
-	"errors"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -13,119 +11,11 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
-	"github.com/ua-academy-projects/share-bite/internal/admin-auth/dto"
 	apperr "github.com/ua-academy-projects/share-bite/internal/admin-auth/error"
-	"github.com/ua-academy-projects/share-bite/internal/admin-auth/handler/auth" // пакет з твоїми хендлерами та Request-структурами
-	"github.com/ua-academy-projects/share-bite/internal/admin-auth/models"
+	"github.com/ua-academy-projects/share-bite/internal/admin-auth/handler/auth"         // пакет з твоїми хендлерами та Request-структурами
 	authsvc "github.com/ua-academy-projects/share-bite/internal/admin-auth/service/auth" // пакет з інтерфейсами провайдера та сервісу
 	"github.com/ua-academy-projects/share-bite/internal/middleware"
 )
-
-type MockProviderFactory struct {
-	mock.Mock
-}
-
-func (m *MockProviderFactory) Get(name string) (authsvc.OAuthProvider, error) {
-	args := m.Called(name)
-	if args.Get(0) != nil {
-		return args.Get(0).(authsvc.OAuthProvider), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-type MockOAuthProvider struct {
-	mock.Mock
-}
-
-func (m *MockOAuthProvider) ExchangeCode(ctx context.Context, code string) (*dto.OAuthUserInfo, error) {
-	args := m.Called(ctx, code)
-	if args.Get(0) != nil {
-		return args.Get(0).(*dto.OAuthUserInfo), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-type MockAuthService struct {
-	mock.Mock
-}
-
-func (m *MockAuthService) Login(ctx context.Context, email string, password string) (*authsvc.Tokens, error) {
-	args := m.Called(ctx, email, password)
-	if args.Get(0) != nil {
-		return args.Get(0).(*authsvc.Tokens), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-func (m *MockAuthService) Register(ctx context.Context, email string, password string, slug string) (*authsvc.Tokens, error) {
-	args := m.Called(ctx, email, password, slug)
-	if args.Get(0) != nil {
-		return args.Get(0).(*authsvc.Tokens), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-func (m *MockAuthService) Refresh(ctx context.Context, refreshToken string) (*authsvc.Tokens, error) {
-	args := m.Called(ctx, refreshToken)
-	if args.Get(0) != nil {
-		return args.Get(0).(*authsvc.Tokens), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-func (m *MockAuthService) OAuthLogin(ctx context.Context, provider authsvc.OAuthProvider, code string, slug string) (*authsvc.Tokens, error) {
-	args := m.Called(ctx, provider, code, slug)
-	if args.Get(0) != nil {
-		return args.Get(0).(*authsvc.Tokens), args.Error(1)
-	}
-	return nil, args.Error(1)
-}
-
-func (m *MockAuthService) LinkProvider(ctx context.Context, userID string, provider authsvc.OAuthProvider, code string) error {
-	args := m.Called(ctx, userID, provider, code)
-	return args.Error(0)
-}
-
-func (m *MockAuthService) RecoverAccess(ctx context.Context, email string) error {
-	args := m.Called(ctx, email)
-	return args.Error(0)
-}
-
-func (m *MockAuthService) ResetPassword(ctx context.Context, token string, newPassword string) error {
-	args := m.Called(ctx, token, newPassword)
-	return args.Error(0)
-}
-
-func (m *MockAuthService) GetUserStatus(ctx context.Context, requesterUserID, requesterRole, targetUserID string) (models.UserStatus, error) {
-	args := m.Called(ctx, requesterUserID, requesterRole, targetUserID)
-	if args.Get(0) == nil {
-		return "", args.Error(1)
-	}
-
-	return args.Get(0).(models.UserStatus), args.Error(1)
-}
-
-func (m *MockAuthService) UpdateUserStatus(ctx context.Context, requesterUserID, requesterRole, targetUserID string, status models.UserStatus) error {
-	args := m.Called(ctx, requesterUserID, requesterRole, targetUserID, status)
-	return args.Error(0)
-}
-
-func testErrorMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		c.Next()
-		err := c.Errors.Last()
-		if err == nil {
-			return
-		}
-
-		var appErr *apperr.AppError
-		if errors.As(err.Err, &appErr) {
-			c.JSON(appErr.HTTPStatus(), gin.H{"error": appErr.Message})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "internal server error"})
-	}
-}
 
 func TestHandler_OAuthCallback(t *testing.T) {
 	gin.SetMode(gin.TestMode)
