+++ internal/admin-auth/handler/tests/handler_recover_reset_test.go
@@ -1,159 +1,18 @@
 package tests
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"net/http"
-	"net/http/httptest"
 	"strings"
 	"testing"
 
-	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/mock"
 	"github.com/ua-academy-projects/share-bite/internal/admin-auth/dto"
-	"github.com/ua-academy-projects/share-bite/internal/admin-auth/handler/auth"
 	"github.com/ua-academy-projects/share-bite/internal/admin-auth/models"
 	"github.com/ua-academy-projects/share-bite/internal/admin-auth/pkg"
-	"github.com/ua-academy-projects/share-bite/internal/admin-auth/repository/user"
-	authsvc "github.com/ua-academy-projects/share-bite/internal/admin-auth/service/auth"
-	"github.com/ua-academy-projects/share-bite/pkg/database"
 	"golang.org/x/crypto/bcrypt"
 )
-
-type mockUserRepository struct {
-	mock.Mock
-}
-
-func (m *mockUserRepository) FindByID(ctx context.Context, userID string) (*dto.UserWithRole, error) {
-	args := m.Called(ctx, userID)
-	userWithRole := args.Get(0)
-	if userWithRole == nil {
-		return nil, args.Error(1)
-	}
-
-	return userWithRole.(*dto.UserWithRole), args.Error(1)
-}
-
-func (m *mockUserRepository) FindByEmail(ctx context.Context, email string) (*dto.UserWithRole, error) {
-	args := m.Called(ctx, email)
-	userWithRole := args.Get(0)
-	if userWithRole == nil {
-		return nil, args.Error(1)
-	}
-
-	return userWithRole.(*dto.UserWithRole), args.Error(1)
-}
-
-func (m *mockUserRepository) FindRoleBySlug(ctx context.Context, slug string) (*models.Role, error) {
-	args := m.Called(ctx, slug)
-	role := args.Get(0)
-	if role == nil {
-		return nil, args.Error(1)
-	}
-
-	return role.(*models.Role), args.Error(1)
-}
-
-func (m *mockUserRepository) CreateUser(ctx context.Context, params user.CreateUser) (*user.CreatedUser, error) {
-	args := m.Called(ctx, params)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-
-	return args.Get(0).(*user.CreatedUser), args.Error(1)
-}
-
-func (m *mockUserRepository) FindBySocialProvider(ctx context.Context, provider string, providerID string) (*dto.UserWithRole, error) {
-	args := m.Called(ctx, provider, providerID)
-	userWithRole := args.Get(0)
-	if userWithRole == nil {
-		return nil, args.Error(1)
-	}
-	return userWithRole.(*dto.UserWithRole), args.Error(1)
-}
-
-func (m *mockUserRepository) CreateWithSocial(ctx context.Context, params dto.CreateUserWithSocialParams) (*dto.CreatedUser, error) {
-	args := m.Called(ctx, params)
-	createdUser := args.Get(0)
-	if createdUser == nil {
-		return nil, args.Error(1)
-	}
-	return createdUser.(*dto.CreatedUser), args.Error(1)
-}
-
-func (m *mockUserRepository) LinkSocialAccount(ctx context.Context, params dto.CreateSocialAccountParams) error {
-	args := m.Called(ctx, params)
-	return args.Error(0)
-}
-
-func (m *mockUserRepository) CreatePasswordResetToken(ctx context.Context, params dto.CreatePasswordResetTokenParams) error {
-	args := m.Called(ctx, params)
-	return args.Error(0)
-}
-
-func (m *mockUserRepository) ResetPassword(ctx context.Context, tokenHash, passwordHash string) (bool, error) {
-	args := m.Called(ctx, tokenHash, passwordHash)
-	return args.Bool(0), args.Error(1)
-}
-
-func (m *mockUserRepository) UpdateUserStatus(ctx context.Context, params user.UpdateUserStatus) error {
-	args := m.Called(ctx, params)
-	return args.Error(0)
-}
-
-func (m *mockUserRepository) AssignRole(ctx context.Context, userID string, roleID int) error {
-	args := m.Called(ctx, userID, roleID)
-	return args.Error(0)
-}
-
-type stubTokenProvider struct{}
-
-func (s stubTokenProvider) GenerateToken(_ string, _ string) (string, string, error) {
-	return "", "", errors.New("unexpected GenerateToken call")
-}
-
-func (s stubTokenProvider) ParseRefreshToken(_ string) (string, string, error) {
-	return "", "", errors.New("unexpected ParseRefreshToken call")
-}
-
-type mockEmailSender struct {
-	mock.Mock
-}
-
-func (s *mockEmailSender) SendPasswordResetToken(ctx context.Context, toEmail, token string) error {
-	args := s.Called(ctx, toEmail, token)
-	return args.Error(0)
-}
-
-type noopTxManager struct{}
-
-func (noopTxManager) ReadCommitted(ctx context.Context, fn database.Handler) error {
-	return fn(ctx)
-}
-
-func buildRecoverResetHandler(repo *mockUserRepository, emailSender *mockEmailSender) *auth.Handler {
-	service := authsvc.New(repo, stubTokenProvider{}, emailSender, noopTxManager{})
-	return auth.NewHandler(service, nil)
-}
-
-func buildGinContext(t *testing.T, method, target string, body any) (*gin.Context, *httptest.ResponseRecorder) {
-	t.Helper()
-
-	gin.SetMode(gin.TestMode)
-	w := httptest.NewRecorder()
-	c, _ := gin.CreateTestContext(w)
-
-	payload, err := json.Marshal(body)
-	if err != nil {
-		t.Fatalf("marshal body: %v", err)
-	}
-
-	c.Request = httptest.NewRequest(method, target, strings.NewReader(string(payload)))
-	c.Request.Header.Set("Content-Type", "application/json")
-
-	return c, w
-}
 
 func TestRecoverAccessSuccess(t *testing.T) {
 	userEmail := "admin@example.com"
