+++ internal/admin-auth/handler/tests/mocks_test.go
@@ -93,6 +93,11 @@ func (m *mockUserRepository) LinkSocialAccount(ctx context.Context, params dto.C
 	return args.Error(0)
 }
 
+func (m *mockUserRepository) UpdateUserStatus(ctx context.Context, params user.UpdateUserStatus) error {
+	args := m.Called(ctx, params)
+	return args.Error(0)
+}
+
 func (m *mockUserRepository) CreatePasswordResetToken(ctx context.Context, params dto.CreatePasswordResetTokenParams) error {
 	args := m.Called(ctx, params)
 	return args.Error(0)
@@ -174,6 +179,19 @@ func (m *MockAuthService) OAuthLogin(ctx context.Context, provider authsvc.OAuth
 
 func (m *MockAuthService) LinkProvider(ctx context.Context, userID string, provider authsvc.OAuthProvider, code string) error {
 	args := m.Called(ctx, userID, provider, code)
+	return args.Error(0)
+}
+
+func (m *MockAuthService) GetUserStatus(ctx context.Context, requesterUserID, requesterRole, targetUserID string) (models.UserStatus, error) {
+	args := m.Called(ctx, requesterUserID, requesterRole, targetUserID)
+	if args.Get(0) != nil {
+		return args.Get(0).(models.UserStatus), args.Error(1)
+	}
+	return "", args.Error(1)
+}
+
+func (m *MockAuthService) UpdateUserStatus(ctx context.Context, requesterUserID, requesterRole, targetUserID string, status models.UserStatus) error {
+	args := m.Called(ctx, requesterUserID, requesterRole, targetUserID, status)
 	return args.Error(0)
 }
